Clean up code, add some error handling

+++ PurchaseTracker/Controllers/HomeController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
+using Kendo.Mvc.UI;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Primitives;
 using Newtonsoft.Json;
@@ -15,31 +16,37 @@ namespace PurchaseTracker.Controllers
         // Database context
         private readonly PurchasesDbContext _context;
 
-        // Category list
-        private Dictionary<int, string> _categories;
+        // Category list caches
+        private Dictionary<int, string> _categoryMap;
+        private List<CategoryViewModel> _categoryViewModels;
 
         public HomeController(PurchasesDbContext context)
         {
             _context = context;
 
-            // Maintain a cache of the category list so we don't have to hit the
-            // database every time we need to look up a category name
-            _categories = new Dictionary<int, string>();
+            // Maintain a cache of the category list so we don't have to
+            // hit the database every time we need to look one up
+            _categoryMap = new Dictionary<int, string>();
+            _categoryViewModels = new List<CategoryViewModel>();
             foreach (Category category in _context.Categories)
             {
-                _categories.Add(category.Id, category.Name);
+                _categoryMap.Add(category.Id, category.Name);
+                _categoryViewModels.Add(new CategoryViewModel() {
+                    Id = category.Id,
+                    Name = category.Name
+                });
             }
         }
 
         // Index view
-        public IActionResult Index()
+        public ActionResult Index()
         {
             return View();
         }
 
         // Error view
         [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
-        public IActionResult Error()
+        public ActionResult Error()
         {
             return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
         }
@@ -48,12 +55,7 @@ namespace PurchaseTracker.Controllers
         [HttpGet]
         public ActionResult GetCategories()
         {
-            var categoryList = new List<Object>();
-            foreach (KeyValuePair<int, string> category in _categories)
-            {
-                categoryList.Add(new { Id = category.Key, Name = category.Value });
-            }
-            return Json(categoryList);
+            return Json(_categoryViewModels);
         }
 
 
@@ -61,20 +63,26 @@ namespace PurchaseTracker.Controllers
         [HttpGet]
         public ActionResult GetPurchases()
         {
-            List<Purchase> purchases = _context.Purchases.ToList();
             List<PurchaseViewModel> viewModels = new List<PurchaseViewModel>();
-            foreach(var p in purchases)
+            try
             {
-                viewModels.Add(new PurchaseViewModel()
+                foreach (Purchase purchase in _context.Purchases)
                 {
-                    Id = p.Id,
-                    CategoryId = p.CategoryId,
-                    CategoryName = _categories.GetValueOrDefault(p.CategoryId),
-                    Payee = p.Payee,
-                    Amount = p.Amount,
-                    Date = p.Date,
-                    Memo = p.Memo
-                });
+                    viewModels.Add(new PurchaseViewModel()
+                    {
+                        Id = purchase.Id,
+                        CategoryId = purchase.CategoryId,
+                        CategoryName = _categoryMap.GetValueOrDefault(purchase.CategoryId),
+                        Payee = purchase.Payee,
+                        Amount = purchase.Amount,
+                        Date = purchase.Date,
+                        Memo = purchase.Memo
+                    });
+                }
+            }
+            catch
+            {
+                return Json(new DataSourceResult() { Errors = "Could not connect to the database." });
             }
             return Json(viewModels);
         }
@@ -83,10 +91,22 @@ namespace PurchaseTracker.Controllers
         [HttpPost]
         public ActionResult UpdatePurchase()
         {
-            PurchaseViewModel model = GetPurchaseViewModel();
-            Purchase purchase = _context.Purchases.Find(model.Id);
-            UpdatePurchase(ref purchase, ref model);
-            _context.SaveChanges();
+            if (!TryGetPurchaseViewModel(out PurchaseViewModel model))
+            {
+                return Json(new DataSourceResult() { Errors = "Model could not be deserialized." });
+            }
+
+            try
+            {
+                Purchase purchase = _context.Purchases.Find(model.Id);
+                UpdatePurchaseModel(ref purchase, ref model);
+                _context.SaveChanges();
+            }
+            catch(Exception)
+            {
+                return Json(new DataSourceResult() { Errors = "Could not connect to the database." });
+            }
+
             return Json(model);
         }
 
@@ -94,10 +114,22 @@ namespace PurchaseTracker.Controllers
         [HttpPost]
         public ActionResult DestroyPurchase()
         {
-            PurchaseViewModel model = GetPurchaseViewModel();
-            Purchase purchase = _context.Purchases.Find(model.Id);
-            _context.Purchases.Remove(purchase);
-            _context.SaveChanges();
+            if (!TryGetPurchaseViewModel(out PurchaseViewModel model))
+            {
+                return Json(new DataSourceResult() { Errors = "Model could not be deserialized." });
+            }
+
+            try
+            {
+                Purchase purchase = _context.Purchases.Find(model.Id);
+                _context.Purchases.Remove(purchase);
+                _context.SaveChanges();
+            }
+            catch(Exception)
+            {
+                return Json(new DataSourceResult() { Errors = "Could not connect to the database." });
+            }
+
             return Json(model);
         }
 
@@ -105,30 +137,51 @@ namespace PurchaseTracker.Controllers
         [HttpPost]
         public ActionResult CreatePurchase()
         {
-            PurchaseViewModel model = GetPurchaseViewModel();
-            Purchase purchase = new Purchase();
-            UpdatePurchase(ref purchase, ref model);
-            _context.Purchases.Add(purchase);
-            _context.SaveChanges();
+            if (!TryGetPurchaseViewModel(out PurchaseViewModel model))
+            {
+                return Json(new DataSourceResult() { Errors = "Model could not be deserialized." });
+            }
+            try
+            {
+                Purchase purchase = new Purchase();
+                UpdatePurchaseModel(ref purchase, ref model);
+                _context.Purchases.Add(purchase);
+                _context.SaveChanges();
+                model.Id = purchase.Id;
+            }
+            catch(Exception)
+            {
+                return Json(new DataSourceResult() { Errors = "Could not connect to the database." });
+            }
+
             return Json(model);
         }
 
         // Deserialize the request string to an instance of PurchaseViewModel
-        private PurchaseViewModel GetPurchaseViewModel()
+        private bool TryGetPurchaseViewModel(out PurchaseViewModel model)
         {
-            StringValues modelsString = Request.Form["models"];
-            return JsonConvert.DeserializeObject<List<PurchaseViewModel>>(modelsString).First();
+            try
+            {
+                StringValues modelsString = Request.Form["models"];
+                model = JsonConvert.DeserializeObject<List<PurchaseViewModel>>(modelsString).First();
+            }
+            catch(Exception)
+            {
+                model = new PurchaseViewModel();
+                return false;
+            }
+            return true;
         }
 
         // Update a Purchase from a PurchaseViewModel
-        private void UpdatePurchase(ref Purchase p, ref PurchaseViewModel model)
+        private void UpdatePurchaseModel(ref Purchase p, ref PurchaseViewModel model)
         {
             p.CategoryId = model.CategoryId;
             p.Payee = model.Payee;
             p.Amount = model.Amount;
             p.Memo = model.Memo;
             p.Date = model.Date;
-            model.CategoryName = _categories.GetValueOrDefault(model.CategoryId);
+            model.CategoryName = _categoryMap.GetValueOrDefault(model.CategoryId);
         }
 
     }
